[Transaction] Rollback previous transaction before loading data
If a load of test data was requested, first check if previous
transaction was closed and rollback it if needed. It prevents
an error when user's test execution was interrupted and data
rollback didn't get requested. Moreover, it enables the user to
skip the rollback request altogether and just expect that a next
load request will handle the clear up.